requirements.txt
pytest == 7.4.2
selenium == 4.15.2
webdriver-manager == 4.0.1
pytest-html == 4.1.1
pytest-xdist == 3.3.1
python-dotenv == 1.2.1